refactor: move theme dir to .dumi/theme
