+++ site/.dumi/theme/examples/bar-aggregated.ts
