+++ site/.dumi/theme/examples/bar-range.ts
