+++ site/.dumi/theme/examples/rose-label.ts
