+++ site/.dumi/theme/utils/copyToBoard.ts
